sdk v1.1.12 remove deprecated productCodes
