+++ pom.xml
@@ -4,7 +4,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.logicommerce</groupId>
 	<artifactId>sdktest</artifactId>
-	<version>1.1.6</version>
+	<version>1.1.7</version>
 	<packaging>jar</packaging>
 	<name>${project.groupId}:${project.artifactId}</name>
 	<description>Logicommerce SDK Testing tools for Logicommerce plugins.</description>
@@ -50,7 +50,7 @@
 		<maven-source-plugin.version>3.2.1</maven-source-plugin.version>
 		<maven-javadoc-plugin.version>3.3.0</maven-javadoc-plugin.version>
 		<nexus-staging-maven-plugin.version>1.6.8</nexus-staging-maven-plugin.version>
-		<logicommerce-sdk.version>1.1.10</logicommerce-sdk.version>
+		<logicommerce-sdk.version>1.1.12</logicommerce-sdk.version>
 		<logicommerce-utilities.version>1.0.3</logicommerce-utilities.version>
 		<apache-httpclient5-version>5.1.3</apache-httpclient5-version>
 		<asynchttpclient.version>2.12.3</asynchttpclient.version>
