+++ src/main/java/com/logicommerce/sdktest/models/RowCodesFakeBuilder.java
@@ -1,88 +1,26 @@
 package com.logicommerce.sdktest.models;
 
-import com.logicommerce.sdk.models.RowCodes;
+import com.logicommerce.sdk.builders.RowCodesBuilder;
 
-public class RowCodesFakeBuilder {
-
-	private CartItemFakeBuilder parentBuilder;
-
-	private String pId;
-
-	private String sku;
-
-	private String ean;
-
-	private String isbn;
-
-	private String jan;
-
-	private String upc;
-	
-	private String manufacturerSku;
+public class RowCodesFakeBuilder<T> extends RowCodesBuilder<T> {
 
 	public RowCodesFakeBuilder() {
-		pId = "default_pId";
-		sku = "default_sku";
-		ean = "default_ean";
-		isbn = "default_isbn";
-		jan = "default_jan";
-		upc = "default_upc";
-		manufacturerSku = "default_manufacturerSku";
+		super();
+		defaultValues();
 	}
 
-	public RowCodesFakeBuilder(CartItemFakeBuilder parentBuilder) {
-		this();
-		this.parentBuilder = parentBuilder;
+	public RowCodesFakeBuilder(T parentBuilder) {
+		super(parentBuilder);
+		defaultValues();
 	}
 
-	public RowCodesFakeBuilder pId(String pId) {
-		this.pId = pId;
-		return this;
-	}
-
-	public RowCodesFakeBuilder sku(String sku) {
-		this.sku = sku;
-		return this;
-	}
-
-	public RowCodesFakeBuilder ean(String ean) {
-		this.ean = ean;
-		return this;
-	}
-
-	public RowCodesFakeBuilder isbn(String isbn) {
-		this.isbn = isbn;
-		return this;
-	}
-
-	public RowCodesFakeBuilder jan(String jan) {
-		this.jan = jan;
-		return this;
-	}
-
-	public RowCodesFakeBuilder upc(String upc) {
-		this.upc = upc;
-		return this;
-	}
-
-	public RowCodesFakeBuilder manufacturerSku(String manufacturerSku) {
-		this.manufacturerSku = manufacturerSku;
-		return this;
-	}
-
-	public RowCodes build() {
-		RowCodesFake codes = new RowCodesFake();
-		codes.setpId(pId);
-		codes.setSku(sku);
-		codes.setEan(ean);
-		codes.setIsbn(isbn);
-		codes.setJan(jan);
-		codes.setUpc(upc);
-		codes.setManufacturerSku(manufacturerSku);
-		return codes;
-	}
-
-	public CartItemFakeBuilder done() {
-		return parentBuilder;
+	private void defaultValues() {
+		super.pId = "default_pId";
+		super.ean = "ean DEFAULT";
+		super.isbn = "isbn DEFAULT";
+		super.jan = "jan DEFAULT";
+		super.manufacturerSku = "manufacturerSku DEFAULT";
+		super.sku = "sku DEFAULT";
+		super.upc = "upc DEFAULT";		
 	}
 }
