+++ src/main/java/com/logicommerce/sdktest/models/order/OrderItemFake.java
@@ -3,7 +3,6 @@ package com.logicommerce.sdktest.models.order;
 import java.util.List;
 import com.logicommerce.sdk.enums.BackorderMode;
 import com.logicommerce.sdk.models.CustomTag;
-import com.logicommerce.sdk.models.ProductCodes;
 import com.logicommerce.sdk.models.RowCodes;
 import com.logicommerce.sdk.models.order.OrderDiscount;
 import com.logicommerce.sdk.models.order.OrderItem;
@@ -31,8 +30,7 @@ public class OrderItemFake implements OrderItem {
 	private boolean sale;
 	private boolean stockManagement;
 	private boolean reverseChargeVat;
-	private ProductCodes codes;
-	private RowCodes rowCodes;
+	private RowCodes codes;
 	private boolean noReturn;
 	private BackorderMode backOrder;
 	private boolean onRequest;
@@ -44,11 +42,6 @@ public class OrderItemFake implements OrderItem {
 	@Override
 	public BackorderMode getBackOrder() {
 		return backOrder;
-	}
-
-	@Override
-	public ProductCodes getCodes() {
-		return codes;
 	}
 
 	@Override
@@ -160,10 +153,6 @@ public class OrderItemFake implements OrderItem {
 		this.backOrder = backOrder;
 	}
 
-	public void setCodes(ProductCodes codes) {
-		this.codes = codes;
-	}
-
 	public void setCustomTags(List<CustomTag> customTags) {
 		this.customTags = customTags;
 	}
@@ -268,11 +257,11 @@ public class OrderItemFake implements OrderItem {
 	}
 
 	@Override
-	public RowCodes getRowCodes() {
-		return rowCodes;
+	public RowCodes getCodes() {
+		return codes;
 	}
 
-	public void setRowCodes(RowCodes rowCodes) {
-		this.rowCodes = rowCodes;
+	public void setCodes(RowCodes codes) {
+		this.codes = codes;
 	}
 }
